requirements.txt
gradio
numpy
Pillow
torch
torchvision
matplotlib
opencv-python